refactor: Introduce constants for URL schemes and service statuses
- Added constants for HTTP and HTTPS schemes to improve URL validation readability.
- Introduced service status and alert level constants to enhance clarity in health metrics handling.
- Updated relevant code sections to utilize these constants, ensuring consistency and maintainability across the codebase.

+++ pre-processor/app/service/health_metrics_test.go
@@ -391,7 +391,8 @@ func TestHealthMetricsCollector_ConcurrentAccess(t *testing.T) {
 
 	// Verify final metrics
 	metrics := collector.GetHealthMetrics(ctx)
-	expectedRequests := uint64(numWorkers * requestsPerWorker)
+	// Convert to uint64 before multiplication to prevent overflow
+	expectedRequests := uint64(numWorkers) * uint64(requestsPerWorker)
 
 	assert.Equal(t, expectedRequests, metrics.RequestCount,
 		"should record all concurrent requests")
